// libs/bookstore/data-access/src/generated/angular/model/book.model.ts
/**
 * BookStoreBff
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The default book object
 */
export interface Book { 
    /**
     * Unique id
     */
    id: string;
    /**
     * Title
     */
    title: string;
    /**
     * Price for this book
     */
    price?: number;
    /**
     * Is book on sale
     */
    onSale?: boolean;
    /**
     * Number of pages
     */
    pageCount?: number;
    /**
     * Last updated is used for concurrency check
     */
    lastUpdated?: string;
    /**
     * last updated by in db
     */
    lastUpdatedBy?: string;
}